adding UserId to Workouts and Completed Workouts

+++ WorkoutBackend.Data/Sql/CompletedExerciseGroups/DeleteCompletedExerciseGroupById.sql
@@ -1,2 +1,3 @@
 DELETE FROM CompletedExerciseGroups
-WHERE Id = @Id;
+WHERE Id = @Id
+	AND UserId = @UserId;
