+++ WorkoutBackend.Data/Sql/CompletedWorkouts/DeleteCompletedWorkoutById.sql
@@ -1,2 +1,3 @@
 DELETE FROM CompletedWorkouts
-WHERE Id = @Id;
+WHERE Id = @Id
+	AND UserId = @UserId;
