+++ WorkoutBackend.Data/Sql/ExerciseGroups/DeleteExerciseGroupById.sql
@@ -1,2 +1,3 @@
 DELETE FROM ExerciseGroups
 WHERE Id = @Id
+	AND UserId = @UserId;
