+++ WorkoutBackend.Data/Sql/ExerciseSets/DeleteExerciseSetById.sql
@@ -1,2 +1,3 @@
 DELETE FROM ExerciseSets
 WHERE Id = @Id
+	AND UserId = @UserId;
