+++ WorkoutBackend.Data/Sql/Workouts/DeleteWorkoutById.sql
@@ -1,2 +1,3 @@
 DELETE FROM Workouts
 WHERE Id = @Id
+	AND UserId = @UserId;
